chore: tsify driver events library
